old-tests/submission/W3C/domhtml/HTMLIFrameElement11.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>iframe</title>

<script type="text/javascript">function loadComplete() { }</script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-iframe-element.html#dom-iframe-contentdocument" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<iframe id="iframe" src="/support/html/sample.html" width='550' heigh='60'>
</iframe>


<script>
var domtest = async_test("Retrieve the contentDocument attribute of the iframe element     and examine its title.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var testNode;
      var cd;
      var vtitle;
      var doc;
      doc = window.document;
      testNode = doc.getElementById("iframe");
      cd = testNode.contentDocument;
      vtitle = cd.title;
      assert_equals(vtitle,"Sample");
       
    domtest.done();
}));

</script>
<p>Retrieve the contentDocument attribute of the iframe element
    and examine its title.</p>
<div id="log"></div>
</body></html>
